requirements.txt
fastapi==0.110.0
uvicorn==0.34.0
aiosqlite==0.20.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
pyotp==2.9.0
python-multipart==0.0.9
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic[email]==2.6.3
httpx==0.27.0
pytest==7.4.4
pytest-asyncio==0.21.1 
emoji==2.14.0
PyJWT==2.10.1
websockets==14.1 